requirements.txt
flask>=2.0.1
cassandra-driver>=3.25.0
pandas>=1.4.4
numpy>=1.24.0
scikit-learn>=1.0.2
xgboost>=1.5.0
lightgbm>=3.3.2
catboost>=1.0.6
mlflow>=1.26.1
optuna>=2.10.0
shap>=0.40.0
python-dotenv>=0.19.2
gunicorn>=20.1.0
pytest>=7.1.1
black>=22.3.0
flake8>=4.0.1
matplotlib>=3.5.2
seaborn>=0.11.2
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
requests>=2.26.0
joblib>=1.0.1 